requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cloudinary==1.23.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-adminlte3==0.1.6
djangorestframework==3.12.2
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
